re-arrange the package layout
