+++ refactordoc/tests/test_class_doc.py
