+++ refactordoc/util.py
